+++ Plugin/Plugin.cs
@@ -1,230 +1,59 @@
-using System;
 using EnvDTE;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
+using System;
 
 namespace EditorConfig.VisualStudio
 {
+    using Settings;
+
     /// <summary>
     /// This plugin attaches to an editor instance and updates its settings at
     /// the appropriate times
     /// </summary>
-    internal class Plugin
+    internal class Plugin : IDisposable
     {
-        IWpfTextView view;
-        DTE dte;
-        ErrorListProvider messageList;
-        ErrorTask message;
-        Results settings;
+        private readonly IWpfTextView _view;
+        private readonly ITextDocument _doc;
+        private readonly Results _settings;
+        private readonly Loader _loader;
+        private readonly GlobalSettings _globalSettings;
 
-        public Plugin(IWpfTextView view, ITextDocument document, DTE dte, ErrorListProvider messageList)
+        public Plugin(IWpfTextView view, ITextDocument document, DTE app, ErrorListProvider messageList)
         {
-            this.view = view;
-            this.dte = dte;
-            this.messageList = messageList;
-            this.message = null;
+            _view = view;
+            _doc = document;
+            _loader = new Loader(view, document, messageList);
+            _globalSettings = new GlobalSettings(view, app, _settings = _loader.Settings);
 
             document.FileActionOccurred += FileActionOccurred;
-            view.GotAggregateFocus += GotAggregateFocus;
             view.Closed += Closed;
-
-            LoadSettings(document.FilePath);
         }
 
         /// <summary>
         /// Reloads the settings when the filename changes
         /// </summary>
-        void FileActionOccurred(object sender, TextDocumentFileActionEventArgs e)
+        private void FileActionOccurred(object sender, TextDocumentFileActionEventArgs e)
         {
             if ((e.FileActionType & FileActionTypes.DocumentRenamed) != 0)
-                LoadSettings(e.FilePath);
+                _loader.LoadSettings(e.FilePath);
 
-            if (settings != null && view.HasAggregateFocus)
-                ApplyGlobalSettings();
+            if (_settings != null && _view.HasAggregateFocus)
+                _globalSettings.Apply();
         }
 
-        /// <summary>
-        /// Updates the global settings when the local editor receives focus
-        /// </summary>
-        void GotAggregateFocus(object sender, EventArgs e)
+        private void Closed(object sender, EventArgs e)
         {
-            if (settings != null)
-                ApplyGlobalSettings();
+            Dispose();
         }
 
-        /// <summary>
-        /// Removes the any messages when the document is closed
-        /// </summary>
-        void Closed(object sender, EventArgs e)
+        public void Dispose()
         {
-            ClearMessage();
-        }
-
-        /// <summary>
-        /// Loads the settings for the given file path
-        /// </summary>
-        private void LoadSettings(string path)
-        {
-            ClearMessage();
-            settings = null;
-
-            // Prevent parsing of internet-located documents,
-            // or documents that do not have proper paths.
-            if (path.StartsWith("http:", StringComparison.OrdinalIgnoreCase)
-                || path.Equals("Temp.txt"))
-                return;
-
-            try
-            {
-                settings = Core.Parse(path);
-                ApplyLocalSettings();
-            }
-            catch (ParseException e)
-            {
-                ShowError(path, "EditorConfig syntax error in file \"" + e.File + "\", line " + e.Line);
-            }
-            catch (CoreException e)
-            {
-                ShowError(path, "EditorConfig core error: " + e.Message);
-            }
-        }
-
-        /// <summary>
-        /// Applies settings to the local text editor instance
-        /// </summary>
-        private void ApplyLocalSettings()
-        {
-            IEditorOptions options = view.Options;
-            
-            if (settings.ContainsKey("tab_width"))
-            {
-                try
-                {
-                    int value = Convert.ToInt32(settings["tab_width"]);
-                    options.SetOptionValue<int>(DefaultOptions.TabSizeOptionId, value);
-                }
-                catch { }
-            }
-
-            if (settings.ContainsKey("indent_size"))
-            {
-                try
-                {
-                    int value = Convert.ToInt32(settings["indent_size"]);
-                    options.SetOptionValue<int>(DefaultOptions.IndentSizeOptionId, value);
-                }
-                catch { }
-            }
-
-            if (settings.ContainsKey("indent_style"))
-            {
-                string value = settings["indent_style"];
-                if (value == "tab")
-                    options.SetOptionValue<bool>(DefaultOptions.ConvertTabsToSpacesOptionId, false);
-                else if (value == "space")
-                    options.SetOptionValue<bool>(DefaultOptions.ConvertTabsToSpacesOptionId, true);
-            }
-
-            if (settings.ContainsKey("end_of_line"))
-            {
-                string value = settings["end_of_line"];
-                if (value == "lf")
-                {
-                    options.SetOptionValue<string>(DefaultOptions.NewLineCharacterOptionId, "\n");
-                    options.SetOptionValue<bool>(DefaultOptions.ReplicateNewLineCharacterOptionId, false);
-                }
-                else if (value == "cr")
-                {
-                    options.SetOptionValue<string>(DefaultOptions.NewLineCharacterOptionId, "\r");
-                    options.SetOptionValue<bool>(DefaultOptions.ReplicateNewLineCharacterOptionId, false);
-                }
-                else if (value == "crlf")
-                {
-                    options.SetOptionValue<string>(DefaultOptions.NewLineCharacterOptionId, "\r\n");
-                    options.SetOptionValue<bool>(DefaultOptions.ReplicateNewLineCharacterOptionId, false);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Applies settings to the global Visual Studio application. Some
-        /// source-code formatters, such as curly-brace auto-indenter, ignore
-        /// the local text editor settings. This causes horrible bugs when
-        /// the local text-editor settings disagree with the formatter's
-        /// settings. To fix this, just apply the same settings at the global
-        /// application level as well.
-        /// </summary>
-        private void ApplyGlobalSettings()
-        {
-            EnvDTE.Properties props;
-            try
-            {
-                string type = view.TextDataModel.ContentType.TypeName;
-                props = dte.get_Properties("TextEditor", type);
-            }
-            catch
-            {
-                // If the above code didn't work, this particular content type
-                // didn't need its settings changed anyhow
-                return;
-            }
-
-            if (settings.ContainsKey("tab_width"))
-            {
-                try
-                {
-                    int value = Convert.ToInt32(settings["tab_width"]);
-                    props.Item("TabSize").Value = value;
-                }
-                catch { }
-            }
-
-            if (settings.ContainsKey("indent_size"))
-            {
-                try
-                {
-                    int value = Convert.ToInt32(settings["indent_size"]);
-                    props.Item("IndentSize").Value = value;
-                }
-                catch { }
-            }
-
-            if (settings.ContainsKey("indent_style"))
-            {
-                string value = settings["indent_style"];
-                if (value == "tab")
-                    props.Item("InsertTabs").Value = true;
-                else if (value == "space")
-                    props.Item("InsertTabs").Value = false;
-            }
-        }
-
-        /// <summary>
-        /// Adds an error message to the Visual Studio tasks pane
-        /// </summary>
-        void ShowError(string path, string text)
-        {
-            message = new ErrorTask();
-            message.ErrorCategory = TaskErrorCategory.Error;
-            message.Category = TaskCategory.Comments;
-            message.Document = path;
-            message.Line = 0;
-            message.Column = 0;
-            message.Text = text;
-
-            messageList.Tasks.Add(message);
-            messageList.Show();
-        }
-
-        /// <summary>
-        /// Removes the file's messages, if any
-        /// </summary>
-        void ClearMessage()
-        {
-            if (message != null)
-                messageList.Tasks.Remove(message);
-            message = null;
+            _loader.Dispose();
+            _globalSettings.Dispose();
+            _doc.FileActionOccurred -= FileActionOccurred;
+            _view.Closed -= Closed;
         }
     }
 }